+++ web/status-history/src/main/java/net/twisterrob/travel/statushistory/infrastructure/github/contract/GithubCreateIssueRequest.kt
@@ -1,4 +1,4 @@
-package net.twisterrob.travel.statushistory.infrastructure.tickets.github.contract
+package net.twisterrob.travel.statushistory.infrastructure.github.contract
 
 import io.micronaut.serde.annotation.Serdeable
