+++ web/status-history/src/test/java/net/twisterrob/travel/statushistory/infrastructure/feedback/ThrowErrorFeedbackInteractorUnitTest.kt
@@ -1,6 +1,5 @@
-package net.twisterrob.travel.statushistory.infrastructure.tickets.github
+package net.twisterrob.travel.statushistory.infrastructure.feedback
 
-import net.twisterrob.travel.statushistory.infrastructure.feedback.ThrowErrorFeedbackInteractor
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.assertThrows
